[CSI] Handle ASan hooks specially when setting up functions for CSI or CilkSanitizer instrumentation.
If ASan inserts hooks into spawned tasks with other exception-handling logic, it either needs to mark that these hooks do not throw or it needs to invoke these hooks with unwinding blocks that connect with the unwind blocks of the enclosing task.  ASan does not currently do this and instead inserts these hooks naively.  This change works around ASan's behavior by marking that calls to ASan hook in spawned tasks do not throw.